Refactor middleware to use mixin
