+++ config/environments/test.rb
@@ -89,6 +89,6 @@ Rails.application.configure do
   # config.action_view.raise_on_missing_translations = true
 
   # Disable CSS3 and jQuery animations in test mode for speed, consistency and to avoid timing issues.
-  config.middleware.insert_after(Rack::NoAnimations)
+  config.middleware.use Rack::NoAnimations
   config.middleware.insert_after(Rack::NoAnimations, Rack::NoPopups)
 end
