+++ lib/rack/no_popups.rb
@@ -1,34 +1,20 @@
 # Disable alerts during headless chrome run
 #
-class Rack::NoPopups
-  DISABLE_POPUPS_HTML = <<~HTML.freeze
-    <script type="text/javascript">
-      window.alert = function() { };
-      window.confirm = function() { return true; };
-    </script>
-  HTML
+module Rack
+  class NoPopups
+    include Injectable
 
-  def initialize(app)
-    @app = app
-  end
+    DISABLE_POPUPS_HTML = <<~HTML.freeze
+      <script type="text/javascript">
+        window.alert = function() { };
+        window.confirm = function() { return true; };
+      </script>
+    HTML
 
-  def call(env)
-    status, headers, body = @app.call(env)
-    return status, headers, body unless html?(headers)
-    response = Rack::Response.new([], status, headers)
+    private
 
-    body.each { |fragment| response.write inject(fragment) }
-    body.close if body.respond_to?(:close)
-    response.finish
-  end
-
-  private
-
-  def html?(headers)
-    headers['Content-Type'] =~ /html/
-  end
-
-  def inject(fragment)
-    fragment.gsub('</body>', DISABLE_POPUPS_HTML + '</body>')
+    def fragment
+      DISABLE_POPUPS_HTML
+    end
   end
 end
